Fix plot image ordering in server timeline
