+++ src/ipc.rs
@@ -104,8 +104,6 @@ pub enum WorkerToServerIpcMessage {
         mime_type: String,
         data: String,
         is_new: bool,
-        #[serde(default)]
-        stdout_bytes_before: u64,
     },
     /// Emitted exactly when the backend knows the logical request has consumed all queued input.
     /// No later `ReadlineResult` should follow for that same request.
@@ -146,7 +144,7 @@ pub struct IpcPlotImage {
     pub mime_type: String,
     pub data: String,
     pub is_new: bool,
-    pub stdout_bytes_before: u64,
+    pub readline_results_seen: usize,
 }
 
 #[derive(Default, Clone)]
@@ -305,15 +303,18 @@ impl ServerIpcConnection {
                             mime_type,
                             data,
                             is_new,
-                            stdout_bytes_before,
                         } => {
+                            let readline_results_seen = {
+                                let guard = reader_inbox.lock().unwrap();
+                                guard.readline_result_count as usize
+                            };
                             if let Some(handler) = plot_handler.as_ref() {
                                 handler(IpcPlotImage {
                                     id,
                                     mime_type,
                                     data,
                                     is_new,
-                                    stdout_bytes_before,
+                                    readline_results_seen,
                                 });
                             } else {
                                 let mut guard = reader_inbox.lock().unwrap();
@@ -322,7 +323,6 @@ impl ServerIpcConnection {
                                     mime_type,
                                     data,
                                     is_new,
-                                    stdout_bytes_before,
                                 });
                                 reader_cvar.notify_all();
                             }
@@ -1329,20 +1329,13 @@ pub fn emit_readline_result(prompt: &str, line: &str) {
     }
 }
 
-pub fn emit_plot_image(
-    id: &str,
-    mime_type: &str,
-    data: &str,
-    is_new: bool,
-    stdout_bytes_before: u64,
-) {
+pub fn emit_plot_image(id: &str, mime_type: &str, data: &str, is_new: bool) {
     if let Some(ipc) = global_ipc() {
         let _ = ipc.send(WorkerToServerIpcMessage::PlotImage {
             id: id.to_string(),
             mime_type: mime_type.to_string(),
             data: data.to_string(),
             is_new,
-            stdout_bytes_before,
         });
     }
 }
