+++ src/pending_output_tape.rs
@@ -2,6 +2,10 @@ use std::collections::VecDeque;
 use std::fmt::Write as _;
 use std::sync::{Arc, Mutex};
 
+use crate::ipc::IpcEchoEvent;
+use crate::output_capture::{OutputEvent, OutputEventKind, OutputRange, OutputTextSpan};
+use crate::output_timeline::collapse_echo_with_attribution;
+use crate::pager;
 use crate::worker_protocol::{ContentOrigin, TextStream, WorkerContent};
 
 #[derive(Clone, Default)]
@@ -16,7 +20,6 @@ struct PendingOutputTapeInner {
     events: VecDeque<PendingOutputEvent>,
     stdout_tail: PendingTextTail,
     stderr_tail: PendingTextTail,
-    worker_stdout_bytes: u64,
     pending_echo_prefix: String,
     last_rendered_text: Option<RenderedTextState>,
 }
@@ -26,7 +29,6 @@ struct PendingTextTail {
     bytes: Vec<u8>,
     origin: Option<ContentOrigin>,
     start_seq: Option<u64>,
-    worker_stdout_start: Option<u64>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -37,8 +39,6 @@ pub(crate) enum PendingOutputEvent {
         origin: ContentOrigin,
         bytes: Vec<u8>,
         terminated: bool,
-        worker_stdout_start: Option<u64>,
-        worker_stdout_end: Option<u64>,
     },
     Image {
         seq: u64,
@@ -46,7 +46,7 @@ pub(crate) enum PendingOutputEvent {
         mime_type: String,
         id: String,
         is_new: bool,
-        stdout_bytes_before: u64,
+        readline_results_seen: usize,
     },
     Sideband {
         seq: u64,
@@ -90,6 +90,13 @@ struct RenderedTextState {
     stream: TextStream,
     origin: ContentOrigin,
     terminated: bool,
+}
+
+struct RenderedPendingOutput {
+    range: OutputRange,
+    echo_events: Vec<IpcEchoEvent>,
+    prompt_variants: Vec<String>,
+    saw_stderr: bool,
 }
 
 impl PendingOutputTape {
@@ -169,7 +176,7 @@ impl PendingOutputTape {
         mime_type: String,
         data: String,
         is_new: bool,
-        stdout_bytes_before: u64,
+        readline_results_seen: usize,
     ) {
         let mut guard = self
             .inner
@@ -179,7 +186,7 @@ impl PendingOutputTape {
         flush_tail(&mut guard, TextStream::Stdout, false);
         flush_tail(&mut guard, TextStream::Stderr, false);
         let seq = next_seq(&mut guard);
-        append_image_event(
+        append_event(
             &mut guard,
             PendingOutputEvent::Image {
                 seq,
@@ -187,7 +194,7 @@ impl PendingOutputTape {
                 mime_type,
                 id,
                 is_new,
-                stdout_bytes_before,
+                readline_results_seen,
             },
         );
     }
@@ -294,30 +301,22 @@ impl PendingOutputTape {
         {
             flush_tail(&mut guard, stream, true);
         }
-        let is_worker_stdout =
-            matches!(stream, TextStream::Stdout) && matches!(origin, ContentOrigin::Worker);
         if tail_mut(&mut guard, stream).bytes.is_empty() {
             let seq = next_seq(&mut guard);
-            let worker_stdout_start = is_worker_stdout.then_some(guard.worker_stdout_bytes);
             let tail = tail_mut(&mut guard, stream);
             tail.start_seq = Some(seq);
-            tail.worker_stdout_start = worker_stdout_start;
         }
         let tail = tail_mut(&mut guard, stream);
         if tail.origin.is_none() {
             tail.origin = Some(origin);
         }
         tail.bytes.extend_from_slice(bytes);
-        if is_worker_stdout {
-            guard.worker_stdout_bytes = guard
-                .worker_stdout_bytes
-                .saturating_add(bytes.len().try_into().unwrap_or(u64::MAX));
-        }
         commit_complete_lines(&mut guard, stream);
     }
 }
 
 impl PendingOutputSnapshot {
+    #[cfg(test)]
     pub(crate) fn format_contents(&self) -> FormattedPendingOutput {
         let mut formatted = FormattedPendingOutput::default();
         let mut last_rendered_text = self.prior_rendered_text;
@@ -375,6 +374,124 @@ impl PendingOutputSnapshot {
             &mut formatted.contents,
         );
         formatted
+    }
+
+    pub(crate) fn format_contents_for_reply(&self) -> FormattedPendingOutput {
+        let RenderedPendingOutput {
+            range,
+            echo_events,
+            prompt_variants,
+            saw_stderr,
+        } = self.rendered_output_for_reply();
+        let source_end = range.end_offset;
+        let collapsed = collapse_echo_with_attribution(range, &echo_events, &prompt_variants);
+        let mut contents = pager::contents_from_collapsed_output(
+            collapsed.bytes,
+            collapsed.events,
+            collapsed.text_spans,
+            source_end,
+        );
+        maybe_trim_leading_echo_prefix(self.leading_echo_prefix.as_deref(), &mut contents);
+        FormattedPendingOutput {
+            contents,
+            saw_stderr,
+        }
+    }
+
+    fn rendered_output_for_reply(&self) -> RenderedPendingOutput {
+        let mut bytes = Vec::new();
+        let mut text_spans = Vec::new();
+        let mut events = Vec::new();
+        let mut echo_events = Vec::new();
+        let mut prompt_variants = Vec::new();
+        let mut saw_stderr = false;
+        let mut last_rendered_text = self.prior_rendered_text;
+
+        for event in &self.events {
+            match event {
+                PendingOutputEvent::TextFragment {
+                    stream,
+                    origin,
+                    bytes: fragment,
+                    terminated,
+                    ..
+                } => {
+                    if fragment.is_empty() {
+                        continue;
+                    }
+                    if matches!(stream, TextStream::Stderr) {
+                        saw_stderr = true;
+                    }
+                    let rendered = render_bytes(fragment);
+                    if rendered.is_empty() {
+                        continue;
+                    }
+                    let text = if matches!(stream, TextStream::Stderr) {
+                        render_stderr_text(last_rendered_text, *origin, rendered)
+                    } else {
+                        rendered
+                    };
+                    append_rendered_text(
+                        &mut bytes,
+                        &mut text_spans,
+                        text.as_bytes(),
+                        *stream,
+                        *origin,
+                    );
+                    last_rendered_text = Some(RenderedTextState {
+                        stream: *stream,
+                        origin: *origin,
+                        terminated: *terminated,
+                    });
+                }
+                PendingOutputEvent::Image {
+                    data,
+                    mime_type,
+                    id,
+                    is_new,
+                    readline_results_seen,
+                    ..
+                } => {
+                    events.push(OutputEvent {
+                        offset: bytes.len() as u64,
+                        kind: OutputEventKind::Image {
+                            data: data.clone(),
+                            mime_type: mime_type.clone(),
+                            id: id.clone(),
+                            is_new: *is_new,
+                            readline_results_seen: *readline_results_seen,
+                        },
+                    });
+                    last_rendered_text = None;
+                }
+                PendingOutputEvent::Sideband { kind, .. } => match kind {
+                    PendingSidebandKind::ReadlineStart { prompt } => {
+                        push_prompt_variant(&mut prompt_variants, prompt);
+                    }
+                    PendingSidebandKind::ReadlineResult { prompt, line } => {
+                        push_prompt_variant(&mut prompt_variants, prompt);
+                        echo_events.push(IpcEchoEvent {
+                            prompt: prompt.clone(),
+                            line: line.clone(),
+                        });
+                    }
+                    PendingSidebandKind::RequestEnd | PendingSidebandKind::SessionEnd => {}
+                },
+            }
+        }
+
+        RenderedPendingOutput {
+            range: OutputRange {
+                start_offset: 0,
+                end_offset: bytes.len() as u64,
+                bytes,
+                events,
+                text_spans,
+            },
+            echo_events,
+            prompt_variants,
+            saw_stderr,
+        }
     }
 }
 
@@ -546,6 +663,7 @@ fn common_prefix_len(left: &str, right: &str) -> usize {
     matched
 }
 
+#[cfg(test)]
 fn push_text(
     contents: &mut Vec<WorkerContent>,
     stream: TextStream,
@@ -571,6 +689,43 @@ fn push_text(
         stream,
         origin,
     });
+}
+
+fn append_rendered_text(
+    bytes: &mut Vec<u8>,
+    text_spans: &mut Vec<OutputTextSpan>,
+    text: &[u8],
+    stream: TextStream,
+    origin: ContentOrigin,
+) {
+    if text.is_empty() {
+        return;
+    }
+    let is_stderr = matches!(stream, TextStream::Stderr);
+    let start_byte = bytes.len();
+    bytes.extend_from_slice(text);
+    let end_byte = bytes.len();
+    if let Some(last) = text_spans.last_mut()
+        && last.is_stderr == is_stderr
+        && last.origin == origin
+        && last.end_byte == start_byte
+    {
+        last.end_byte = end_byte;
+    } else {
+        text_spans.push(OutputTextSpan {
+            start_byte,
+            end_byte,
+            is_stderr,
+            origin,
+        });
+    }
+}
+
+fn push_prompt_variant(prompt_variants: &mut Vec<String>, prompt: &str) {
+    if prompt_variants.iter().any(|existing| existing == prompt) {
+        return;
+    }
+    prompt_variants.push(prompt.to_string());
 }
 
 fn render_bytes(bytes: &[u8]) -> String {
@@ -665,15 +820,13 @@ fn append_complete_bytes(
             origin,
             bytes: bytes.to_vec(),
             terminated: bytes.ends_with(b"\n"),
-            worker_stdout_start: None,
-            worker_stdout_end: None,
         },
     );
 }
 
 fn commit_complete_lines(inner: &mut PendingOutputTapeInner, stream: TextStream) {
     loop {
-        let (seq, origin, line, tail_empty, worker_stdout_start, worker_stdout_end) = {
+        let (seq, origin, line, tail_empty) = {
             let tail = tail_mut(inner, stream);
             let Some(newline_idx) = tail.bytes.iter().position(|byte| *byte == b'\n') else {
                 break;
@@ -685,23 +838,12 @@ fn commit_complete_lines(inner: &mut PendingOutputTapeInner, stream: TextStream)
                 .origin
                 .expect("text tail should record origin while bytes are buffered");
             let line = tail.bytes.drain(..=newline_idx).collect::<Vec<u8>>();
-            let worker_stdout_start = tail.worker_stdout_start;
-            let worker_stdout_end = worker_stdout_start
-                .map(|start| start.saturating_add(line.len().try_into().unwrap_or(u64::MAX)));
             let tail_empty = tail.bytes.is_empty();
             if tail_empty {
                 tail.origin = None;
                 tail.start_seq = None;
-                tail.worker_stdout_start = None;
             }
-            (
-                seq,
-                origin,
-                line,
-                tail_empty,
-                worker_stdout_start,
-                worker_stdout_end,
-            )
+            (seq, origin, line, tail_empty)
         };
         append_event(
             inner,
@@ -711,21 +853,18 @@ fn commit_complete_lines(inner: &mut PendingOutputTapeInner, stream: TextStream)
                 origin,
                 bytes: line,
                 terminated: true,
-                worker_stdout_start,
-                worker_stdout_end,
             },
         );
         if !tail_empty {
             let next = next_seq(inner);
             let tail = tail_mut(inner, stream);
             tail.start_seq = Some(next);
-            tail.worker_stdout_start = worker_stdout_end;
         }
     }
 }
 
 fn flush_tail(inner: &mut PendingOutputTapeInner, stream: TextStream, flush_incomplete: bool) {
-    let (seq, origin, bytes, tail_empty, worker_stdout_start, worker_stdout_end) = {
+    let (seq, origin, bytes, tail_empty) = {
         let tail = tail_mut(inner, stream);
         if tail.bytes.is_empty() {
             return;
@@ -744,23 +883,12 @@ fn flush_tail(inner: &mut PendingOutputTapeInner, stream: TextStream, flush_inco
             .origin
             .expect("text tail should record origin while bytes are buffered");
         let bytes = tail.bytes.drain(..flush_len).collect::<Vec<u8>>();
-        let worker_stdout_start = tail.worker_stdout_start;
-        let worker_stdout_end = worker_stdout_start
-            .map(|start| start.saturating_add(bytes.len().try_into().unwrap_or(u64::MAX)));
         let tail_empty = tail.bytes.is_empty();
         if tail_empty {
             tail.origin = None;
             tail.start_seq = None;
-            tail.worker_stdout_start = None;
         }
-        (
-            seq,
-            origin,
-            bytes,
-            tail_empty,
-            worker_stdout_start,
-            worker_stdout_end,
-        )
+        (seq, origin, bytes, tail_empty)
     };
     append_event(
         inner,
@@ -770,15 +898,12 @@ fn flush_tail(inner: &mut PendingOutputTapeInner, stream: TextStream, flush_inco
             origin,
             bytes,
             terminated: false,
-            worker_stdout_start,
-            worker_stdout_end,
         },
     );
     if !tail_empty {
         let next = next_seq(inner);
         let tail = tail_mut(inner, stream);
         tail.start_seq = Some(next);
-        tail.worker_stdout_start = worker_stdout_end;
     }
 }
 
@@ -794,96 +919,6 @@ fn append_event(inner: &mut PendingOutputTapeInner, event: PendingOutputEvent) {
         .position(|existing| existing.seq() > seq)
         .unwrap_or(inner.events.len());
     inner.events.insert(idx, event);
-}
-
-fn append_image_event(inner: &mut PendingOutputTapeInner, event: PendingOutputEvent) {
-    let PendingOutputEvent::Image {
-        stdout_bytes_before,
-        ..
-    } = event
-    else {
-        append_event(inner, event);
-        return;
-    };
-
-    let mut idx = 0usize;
-    while idx < inner.events.len() {
-        let Some(existing) = inner.events.get(idx).cloned() else {
-            break;
-        };
-        let PendingOutputEvent::TextFragment {
-            seq,
-            stream: TextStream::Stdout,
-            origin: ContentOrigin::Worker,
-            bytes,
-            terminated,
-            worker_stdout_start: Some(worker_stdout_start),
-            worker_stdout_end: Some(worker_stdout_end),
-        } = existing
-        else {
-            idx = idx.saturating_add(1);
-            continue;
-        };
-
-        if stdout_bytes_before <= worker_stdout_start {
-            inner.events.insert(idx, event);
-            return;
-        }
-
-        if stdout_bytes_before >= worker_stdout_end {
-            idx = idx.saturating_add(1);
-            continue;
-        }
-
-        let split_at = stdout_bytes_before
-            .saturating_sub(worker_stdout_start)
-            .try_into()
-            .unwrap_or(bytes.len());
-        if !can_split_text_fragment(&bytes, split_at) {
-            idx = idx.saturating_add(1);
-            continue;
-        }
-
-        inner.events.remove(idx);
-        let before_bytes = bytes[..split_at].to_vec();
-        let after_bytes = bytes[split_at..].to_vec();
-        let mut insert_at = idx;
-        if !before_bytes.is_empty() {
-            inner.events.insert(
-                insert_at,
-                PendingOutputEvent::TextFragment {
-                    seq,
-                    stream: TextStream::Stdout,
-                    origin: ContentOrigin::Worker,
-                    bytes: before_bytes,
-                    terminated: false,
-                    worker_stdout_start: Some(worker_stdout_start),
-                    worker_stdout_end: Some(stdout_bytes_before),
-                },
-            );
-            insert_at = insert_at.saturating_add(1);
-        }
-        inner.events.insert(insert_at, event);
-        insert_at = insert_at.saturating_add(1);
-        if !after_bytes.is_empty() {
-            let after_seq = next_seq(inner);
-            inner.events.insert(
-                insert_at,
-                PendingOutputEvent::TextFragment {
-                    seq: after_seq,
-                    stream: TextStream::Stdout,
-                    origin: ContentOrigin::Worker,
-                    bytes: after_bytes,
-                    terminated,
-                    worker_stdout_start: Some(stdout_bytes_before),
-                    worker_stdout_end: Some(worker_stdout_end),
-                },
-            );
-        }
-        return;
-    }
-
-    append_event(inner, event);
 }
 
 fn last_text_fragment_bytes(events: &VecDeque<PendingOutputEvent>) -> Option<&[u8]> {
@@ -946,13 +981,6 @@ fn flushable_prefix_len(bytes: &[u8]) -> usize {
     bytes.len()
 }
 
-fn can_split_text_fragment(bytes: &[u8], split_at: usize) -> bool {
-    split_at == 0
-        || split_at >= bytes.len()
-        || (std::str::from_utf8(&bytes[..split_at]).is_ok()
-            && std::str::from_utf8(&bytes[split_at..]).is_ok())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -973,8 +1001,6 @@ mod tests {
                     origin: ContentOrigin::Worker,
                     bytes: b"abc".to_vec(),
                     terminated: false,
-                    worker_stdout_start: Some(0),
-                    worker_stdout_end: Some(3),
                 },
                 PendingOutputEvent::TextFragment {
                     seq: 1,
@@ -982,8 +1008,6 @@ mod tests {
                     origin: ContentOrigin::Worker,
                     bytes: b"boom\n".to_vec(),
                     terminated: true,
-                    worker_stdout_start: None,
-                    worker_stdout_end: None,
                 },
             ]
         );
@@ -1111,8 +1135,6 @@ mod tests {
                     origin: ContentOrigin::Server,
                     bytes: b"[repl] guardrail".to_vec(),
                     terminated: false,
-                    worker_stdout_start: None,
-                    worker_stdout_end: None,
                 },
                 PendingOutputEvent::TextFragment {
                     seq: 1,
@@ -1120,8 +1142,6 @@ mod tests {
                     origin: ContentOrigin::Worker,
                     bytes: b"ok\n".to_vec(),
                     terminated: true,
-                    worker_stdout_start: Some(0),
-                    worker_stdout_end: Some(3),
                 },
             ]
         );
@@ -1387,8 +1407,6 @@ mod tests {
                     origin: ContentOrigin::Worker,
                     bytes: vec![0xC3],
                     terminated: false,
-                    worker_stdout_start: Some(0),
-                    worker_stdout_end: Some(1),
                 },
                 PendingOutputEvent::TextFragment {
                     seq: 1,
@@ -1396,8 +1414,6 @@ mod tests {
                     origin: ContentOrigin::Server,
                     bytes: b"\n[repl] session ended\n".to_vec(),
                     terminated: true,
-                    worker_stdout_start: None,
-                    worker_stdout_end: None,
                 },
             ]
         );
@@ -1419,8 +1435,6 @@ mod tests {
                     origin: ContentOrigin::Server,
                     bytes: vec![0xC3],
                     terminated: false,
-                    worker_stdout_start: None,
-                    worker_stdout_end: None,
                 },
                 PendingOutputEvent::TextFragment {
                     seq: 1,
@@ -1428,29 +1442,37 @@ mod tests {
                     origin: ContentOrigin::Worker,
                     bytes: b"boom\n".to_vec(),
                     terminated: true,
-                    worker_stdout_start: None,
-                    worker_stdout_end: None,
                 },
             ]
         );
     }
 
     #[test]
-    fn image_anchor_inserts_before_later_stdout_fragment() {
+    fn reply_format_anchors_image_before_later_echoed_input_and_stdout() {
         let tape = PendingOutputTape::new();
 
+        tape.append_stdout_bytes(b"> plot(1:10)\n");
+        tape.append_sideband(PendingSidebandKind::ReadlineResult {
+            prompt: "> ".to_string(),
+            line: "plot(1:10)\n".to_string(),
+        });
+        tape.append_stdout_bytes(b"> cat('done\\n')\n");
         tape.append_stdout_bytes(b"done\n");
         tape.append_image(
             "img-1".to_string(),
             "image/png".to_string(),
             "AA==".to_string(),
             true,
-            0,
+            1,
         );
+        tape.append_sideband(PendingSidebandKind::ReadlineResult {
+            prompt: "> ".to_string(),
+            line: "cat('done\\n')\n".to_string(),
+        });
 
         let snapshot = tape.drain_snapshot();
         assert_eq!(
-            snapshot.format_contents().contents,
+            snapshot.format_contents_for_reply().contents,
             vec![
                 WorkerContent::ContentImage {
                     data: "AA==".to_string(),
